New waypoint movement generator v1
Former-commit-id: 0523b48884b1021d4796a15c01384c7e16012ed7

+++ src/game/Movement/Waypoints/WaypointManager.cpp
@@ -76,7 +76,7 @@ void WaypointMgr::Load()
         wp->move_type = fields[6].GetUInt32();
         wp->delay = fields[7].GetUInt32();
         wp->event_id = fields[8].GetUInt32();
-        wp->event_chance = fields[9].GetInt16();
+        wp->event_chance = fields[9].GetInt8();
 
         path.push_back(wp);
         ++count;
